+++ workflow-fingerprint-collector/CHANGELOG.md
@@ -1,2 +1,4 @@
+## 1.2 - 2015-09-08
+- This upgrade adds an attribute to generated files pointing to the hotspots file used for fingerprinting
 ## 1.1 - 2015-03-19
 - Upgrade to SeqWare 1.1.0, common-utilities 1.6 and workflow-utilities 1.6.